move save in automl.common.util to chronos.feature.utils
