+++ pyzoo/zoo/chronos/feature/utils.py
@@ -52,3 +52,16 @@ def save_config(file_path, config, replace=False):
 
     with open(file_path, "w") as output_file:
         json.dump(config, output_file, cls=NumpyEncoder)
+
+
+def save(file_path, feature_transformers=None, model=None, config=None):
+    if not os.path.isdir(file_path):
+        os.mkdir(file_path)
+    config_path = os.path.join(file_path, "config.json")
+    model_path = os.path.join(file_path, "weights_tune.h5")
+    if feature_transformers is not None:
+        feature_transformers.save(config_path, replace=True)
+    if model is not None:
+        model.save(model_path, config_path)
+    if config is not None:
+        save_config(config_path, config)
